+++ src/components/gcs-layerswitcher/GCSLayerswitcher.js
@@ -7,8 +7,11 @@ export default class GCSLayerSwitcher extends HTMLElement {
         this.control = null;
     }
 
-    registerGCSMap (map, layerManager) {
-        this.control = new LayerswitcherControl(layerManager);
+    registerGCSMap (map, layerManager, i18next) {
+        i18next.addResources("en", "layerswitcher", {OPEN: "Show layerswitcher", CLOSE: "Hide layerswitcher"});
+        i18next.addResources("de", "layerswitcher", {OPEN: "Ebenenauswahl anzeigen", CLOSE: "Ebenenauswahl verbergen"});
+
+        this.control = new LayerswitcherControl(layerManager, i18next);
         map.addControl(this.control);
     }
 }