+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The [Semantic Versioning](https://semver.org/spec/v2.0.0.html) is used.
 ## 1.0.11
 ### Added
 - Added styling for vector layers
+- Language support. DE (default) and EN.
 ### Changed
 ### Deprecated
 ### Removed